feat(skill): introduce SkillGenerator class for format-agnostic skill package compilation
