+++ agent_sdks/python/a2ui_agent/tests/skill/test_skill_generator.py
@@ -12,16 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Unit tests for A2UI Skill and SkillSet composition API."""
+"""Unit tests for A2UI SkillGenerator API."""
 
 import os
-import unittest
 import tempfile
-import yaml
+import unittest
 
 from a2ui.inference_formats.experimental.express import ExpressFormat
 from a2ui.schema.catalog import A2uiCatalog, CatalogConfig
-from a2ui.skill.skill import Skill, SkillSet
+from a2ui.skill import SkillGenerator
 
 
 # Locate standard basic catalog in repository
@@ -40,21 +39,22 @@ SPEC_DIR = os.path.abspath(
 CATALOG_PATH = os.path.join(SPEC_DIR, "catalogs", "basic", "catalog.json")
 
 
-class TestSkillCompositionAPI(unittest.TestCase):
-    """Tests Skill and SkillSet domain composition factory methods."""
+class TestSkillGenerator(unittest.TestCase):
+    """Tests SkillGenerator compilation methods."""
 
     def setUp(self):
         self.temp_dir = tempfile.TemporaryDirectory()
         self.catalog_config = CatalogConfig.from_path("basic", CATALOG_PATH)
         self.catalog = A2uiCatalog.from_config(self.catalog_config)
+        self.express_fmt = ExpressFormat(catalog=self.catalog)
+        self.generator = SkillGenerator(self.express_fmt)
 
     def tearDown(self):
         self.temp_dir.cleanup()
 
-    def test_skill_from_format_monolithic(self):
-        """Verifies Skill.from_format() creating a monolithic Skill object."""
-        express_fmt = ExpressFormat(catalog=self.catalog)
-        skill_obj = Skill.from_format(express_fmt, name="a2ui-custom-monolithic")
+    def test_generate_skill_monolithic(self):
+        """Verifies SkillGenerator.generate_skill() creating a monolithic Skill object."""
+        skill_obj = self.generator.generate_skill(name="a2ui-custom-monolithic")
 
         self.assertEqual(skill_obj.name, "a2ui-custom-monolithic")
         self.assertIn("a2ui-custom-monolithic", skill_obj.filename)
@@ -62,28 +62,25 @@ class TestSkillCompositionAPI(unittest.TestCase):
         self.assertTrue(content.startswith("---"))
         self.assertIn("A2UI Express DSL Output Contract", content)
 
-    def test_skill_from_catalog(self):
-        """Verifies Skill.from_catalog() creating a catalog skill with clean LLM name."""
-        express_fmt = ExpressFormat(catalog=self.catalog)
-        skill_obj = Skill.from_catalog(self.catalog, express_fmt)
+    def test_generate_catalog_skill(self):
+        """Verifies SkillGenerator.generate_catalog_skill() creating a catalog skill with clean LLM name."""
+        skill_obj = self.generator.generate_catalog_skill(self.catalog)
 
         self.assertEqual(skill_obj.name, "a2ui-basic")
         self.assertIn("a2ui-basic/SKILL.md", skill_obj.filename)
         content = skill_obj.to_markdown()
         self.assertIn("Positional Component Signatures", content)
 
-    def test_skill_core_syntax(self):
-        """Verifies Skill.core_syntax() creating base grammar skill."""
-        express_fmt = ExpressFormat(catalog=self.catalog)
-        core_skill = Skill.core_syntax(express_fmt, name="a2ui-base-core")
+    def test_generate_core_skill(self):
+        """Verifies SkillGenerator.generate_core_skill() creating base grammar skill."""
+        core_skill = self.generator.generate_core_skill(name="a2ui-base-core")
 
         self.assertEqual(core_skill.name, "a2ui-base-core")
         self.assertIn("A2UI Express DSL Output Contract", core_skill.content)
 
-    def test_skillset_from_format_modular(self):
-        """Verifies SkillSet.from_format() generating modular skill package."""
-        express_fmt = ExpressFormat(catalog=self.catalog)
-        skill_set = SkillSet.from_format(express_fmt)
+    def test_generate_skillset_modular(self):
+        """Verifies SkillGenerator.generate_skillset() generating modular skill package."""
+        skill_set = self.generator.generate_skillset()
 
         self.assertIn("a2ui-core/SKILL.md", skill_set)
         self.assertIn("a2ui-basic/SKILL.md", skill_set)
@@ -97,8 +94,7 @@ class TestSkillCompositionAPI(unittest.TestCase):
 
     def test_export_to_directory(self):
         """Verifies exporting SkillSet to directory."""
-        express_fmt = ExpressFormat(catalog=self.catalog)
-        skill_set = SkillSet.from_format(express_fmt)
+        skill_set = self.generator.generate_skillset()
         exported = skill_set.export_to_directory(self.temp_dir.name)
 
         self.assertIn("a2ui-core/SKILL.md", exported)
